+++ packages/api/src/handler.ts
@@ -5,29 +5,79 @@ import type {
   Context,
 } from "aws-lambda";
 import {
-  handleHello,
-  handleHelloYou,
-  helloYouDecoder,
-} from "./hello/helloController";
-import { handlePublicGET, handlePublicPOST, setup } from "./services/io/io";
+  adminImageUploadDecoder,
+  adminListDecoder,
+  adminSpaceUpdateDecoder,
+  handleAdminImageUpload,
+  handleAdminSpaceUpdate,
+  handleAdminSpaces,
+} from "./controllers/adminController";
+import {
+  handleGet,
+  handleGetDecoder,
+  handleList,
+} from "./controllers/spaceController";
+import {
+  handleAdminPOST,
+  handlePublicGET,
+  handlePublicPOST,
+  setup,
+} from "./services/io/io";
 
 setup();
 
-export const hello = AWSLambda.wrapHandler(async function hello(
+export const getSpace = AWSLambda.wrapHandler(async function list(
   event: APIGatewayProxyEvent,
   context: Context,
 ): Promise<APIGatewayProxyResult> {
-  return handlePublicGET<"hello">(event, context, handleHello);
+  return handlePublicPOST<"get">(event, context, handleGet, handleGetDecoder);
 });
 
-export const helloYou = AWSLambda.wrapHandler(async function helloYou(
+export const list = AWSLambda.wrapHandler(async function list(
   event: APIGatewayProxyEvent,
   context: Context,
 ): Promise<APIGatewayProxyResult> {
-  return handlePublicPOST<"hello">(
-    event,
-    context,
-    handleHelloYou,
-    helloYouDecoder,
-  );
+  return handlePublicGET<"list">(event, context, handleList);
 });
+
+export const adminImageUpload = AWSLambda.wrapHandler(
+  async function adminImageUpload(
+    event: APIGatewayProxyEvent,
+    context: Context,
+  ): Promise<APIGatewayProxyResult> {
+    return handleAdminPOST<"admin/image/upload">(
+      event,
+      context,
+      handleAdminImageUpload,
+      adminImageUploadDecoder,
+    );
+  },
+);
+
+export const adminSpaceList = AWSLambda.wrapHandler(
+  async function adminSpaceList(
+    event: APIGatewayProxyEvent,
+    context: Context,
+  ): Promise<APIGatewayProxyResult> {
+    return handleAdminPOST<"admin/space">(
+      event,
+      context,
+      handleAdminSpaces,
+      adminListDecoder,
+    );
+  },
+);
+
+export const adminSpaceUpdate = AWSLambda.wrapHandler(
+  async function adminSpaceUpdate(
+    event: APIGatewayProxyEvent,
+    context: Context,
+  ): Promise<APIGatewayProxyResult> {
+    return handleAdminPOST<"admin/space/update">(
+      event,
+      context,
+      handleAdminSpaceUpdate,
+      adminSpaceUpdateDecoder,
+    );
+  },
+);