+++ packages/api/src/hello/message.js
@@ -1,6 +1,0 @@
-// Don't use JavaScript file when possible
-// This file is to test possibility to still use plain js files
-
-export const messages = {
-  missingName: "give your name please",
-};