+++ packages/react/src/context/SpaceServiceSettings.tsx
@@ -1,0 +1,17 @@
+import type { ReactNode } from "react";
+import React from "react";
+import type { ApiSettings } from "../api/api";
+import { useSpaceService } from "../api/useApi";
+
+interface UserServiceSettingsProps {
+  children: JSX.Element | JSX.Element[] | ReactNode;
+  settings: ApiSettings;
+}
+
+export function SpaceServiceSettings(
+  props: UserServiceSettingsProps,
+): JSX.Element {
+  useSpaceService(props.settings);
+  // eslint-disable-next-line react/jsx-no-useless-fragment
+  return <>{props.children}</>;
+}