+++ packages/react/src/index.ts
@@ -1,1 +1,4 @@
-export { Hello } from "./Hello/Hello";
+export { SpaceServiceSettings } from "./context/SpaceServiceSettings";
+export { SpaceList } from "./components/SpaceList";
+export { AdminApp } from "./admin/AdminApp";
+export { useSpace, useSpaceList } from "./api/useApi";